Styled pricing cards
Using Bootstrap classes the cards are set as 3 column on large, 2 on medium with the 3rd taking up 12 and on small defaults to 12 on all 3.

I didn't like the uneven column lengths so I set the row to flex, flex-wrap, set the content to 100% height. I then added flex-column and d-flex to the card body. This still left the buttons uneven so I added mt-auto to the buttons.